+++ tools/themes/tool.js
@@ -8,7 +8,7 @@
         name: 'themes',
         category: 'general',
         type: 'multiple',
-        possibleValues: ['performance', 'justExplorer', 'mobile', 'slimArrows', 'flairX', 'lessIcons', 'nonStickyHeader', 'noStudyChat', 'pieceDrag','noPractice', 'gameMoveList', 'fatGauge'/*, 'noGrab'*/],
+        possibleValues: ['performance', 'justExplorer', 'mobile', 'slimArrows', 'slimmerArrows', 'flairX', 'lessIcons', 'nonStickyHeader', 'noStudyChat', 'pieceDrag','noPractice', 'gameMoveList', 'fatGauge'/*, 'noGrab'*/],
         defaultValue: '',
         advanced: true
       }
@@ -31,7 +31,8 @@
         'themes.noGrab': 'No grab cursor',
         'themes.noPractice': 'No Practice button',
         'themes.gameMoveList': 'Flexible game move list',
-        'themes.fatGauge': 'Thick analysis gauge'
+        'themes.fatGauge': 'Thick analysis gauge',
+        'themes.slimmerArrows': '... slimmer'
       },
       'ro-RO': {
         'options.general': 'General',
@@ -49,7 +50,8 @@
         'themes.noGrab': 'F\u0103r\u0103 cursor care apuc\u0103',
         'themes.noPractice': 'F\u0103r\u0103 buton Antrenament cu calculatorul',
         'themes.gameMoveList': 'List\u0103 mut\u0103ri flexibil\u0103 \u00een joc',
-        'themes.fatGauge': 'Bar\u0103 analiz\u0103 groas\u0103'
+        'themes.fatGauge': 'Bar\u0103 analiz\u0103 groas\u0103',
+        'themes.slimmerArrows': '... \u015fi mai sub\u0163iri'
       }
     }